% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getdata}
\title{getData}
\usage{
getData(input_dir, lines_to_delete = NA, columns_as_num = NA)
}
\arguments{
\item{input_dir}{A character (directory that contains all Sessl results)}

\item{lines_to_delete}{A vector containing all lines that should be
deleted}

\item{columns_as_num}{A vector containing the names (strings) of the
columns to be saven as numeric}
}
\description{
Read csv files from SESSL experiments
}
\details{
Load results from SESSL and save for every directory a RData 
file. The input should be a directory that contains directories with
the config directories and experiment results.
}
\examples{
\dontrun{
}

}
\author{
Kai Budde
}
