% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\alias{plotdata}
\title{plotData}
\usage{
plotData(input_file, output_name, x_axis_name = NA, y_axis_name = NA,
  y_log = FALSE)
}
\arguments{
\item{input_file}{A character (Rda file that contains all Sessl results)}

\item{output_name}{A character (name of the file that is to be saved)}

\item{x_axis_name}{A character}

\item{y_axis_name}{A character}

\item{y_log}{A boolean}
}
\description{
Load data (Rda) of a Sessl experiment and plot it.
}
\details{
TODO
}
\examples{
\dontrun{
}

}
\author{
Kai Budde
}
